pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "qa-z"
version = "0.9.8a0"
description = "A Codex-first QA control plane for coding agents."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [{ name = "QA-Z Contributors" }]
keywords = ["qa", "agents", "codex", "claude", "testing", "ci"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Topic :: Software Development :: Quality Assurance",
  "Topic :: Software Development :: Testing",
]
dependencies = ["PyYAML>=6.0"]

[project.optional-dependencies]
dev = [
  "build>=1.2",
  "mypy>=1.10",
  "pytest>=9.0",
  "ruff>=0.8",
  "types-PyYAML>=6.0",
]

[project.scripts]
qa-z = "qa_z.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
qa_z = ["templates/*.md"]

[tool.ruff]
cache-dir = "~/AppData/Local/Temp/qa-z-ruff-cache"

[tool.pytest.ini_options]
testpaths = ["tests"]
